refactor: cgi 디렉토리 하위 디렉토리 통합
