+++ src/handler/cgi/Executor.cpp
@@ -3,7 +3,7 @@
 
 #include <sys/socket.h>
 
-#include "../../../utils/str_utils.hpp"
+#include "../../utils/str_utils.hpp"
 
 using namespace handler::cgi;
